Reverted. Translated memory allocation in vptree and sptree to C++ standard. Warnings need to be again corrected

+++ datadriven/src/sgpp/datadriven/datamining/modules/visualization/algorithms/bhtsne/tsne.hpp
@@ -50,22 +50,21 @@ class TSNE {
   TSNE();
   void run(double* X, size_t N, size_t D, double* Y, size_t no_dims, double perplexity,
     double theta, size_t rand_seed,
-    bool skip_random_init, size_t max_iter = 1000,  size_t mom_switch_iter = 250);
+    bool skip_random_init, size_t max_iter = 1000,  size_t mom_switch_iter =250);
 
  private:
-  void computeGradient(double* P, size_t* inp_row_P, size_t* inp_col_P,
-    double* inp_val_P, double* Y, size_t N, size_t D, double* dC, double theta);
-  void computeExactGradient(double* P, double* Y, size_t N, size_t D, double* dC);
-  double evaluateError(double* P, double* Y, size_t N, size_t D);
-  double evaluateError(size_t* row_P, size_t* col_P, double* val_P,
-    double* Y, size_t N, size_t D, double theta);
-  void zeroMean(double* X, size_t N, size_t D);
-  void computeGaussianPerplexity(double* X, size_t N, size_t D, double* P, double perplexity);
-  void computeGaussianPerplexity(double* X, size_t N, size_t D, size_t** _row_P,
-    size_t** _col_P, double** _val_P, double perplexity, size_t K);
-  void computeSquaredEuclideanDistance(double* X, size_t N, size_t D, double* DD);
-  void symmetrizeMatrix(size_t** row_P, size_t** col_P, double** val_P,
-    size_t N);
+  void computeGradient(double* P, unsigned int* inp_row_P, unsigned int* inp_col_P,
+    double* inp_val_P, double* Y, int N, int D, double* dC, double theta);
+  void computeExactGradient(double* P, double* Y, int N, int D, double* dC);
+  double evaluateError(double* P, double* Y, int N, int D);
+  double evaluateError(unsigned int* row_P, unsigned int* col_P, double* val_P,
+    double* Y, int N, int D, double theta);
+  void zeroMean(double* X, int N, int D);
+  void computeGaussianPerplexity(double* X, int N, int D, double* P, double perplexity);
+  void computeGaussianPerplexity(double* X, int N, int D, unsigned int** _row_P,
+    unsigned int** _col_P, double** _val_P, double perplexity, int K);
+  void computeSquaredEuclideanDistance(double* X, int N, int D, double* DD);
+  void symmetrizeMatrix(unsigned int** row_P, unsigned int** col_P, double** val_P, int N);
   double randn();
 };
